Switched project to Gradle for hopeful Eclipse compatibility
