+++ .gitignore
@@ -1,13 +1,13 @@
-/src/assignment3/MiniTester.java
-/src/assignment3/CatInfo.java
-/src/assignment3/CatTree.java
-/src/assignment3/listNode.java
-/src/assignment3/bTree.java
-/src/assignment3/Queue.java
-/src/assignment3/bNode.java
-/src/assignment3/Stack.java
-/src/assignment3/DogShelter.java
-/src/assignment3/DogShelter2.java
-/src/assignment3/Treap.java
+/src/test/java/assignment3/MiniTester.java
+/src/test/java/assignment3/CatInfo.java
+/src/test/java/assignment3/CatTree.java
+/src/test/java/assignment3/listNode.java
+/src/test/java/assignment3/bTree.java
+/src/test/java/assignment3/Queue.java
+/src/test/java/assignment3/bNode.java
+/src/test/java/assignment3/Stack.java
+/src/main/java/DogShelter.java
+/src/test/java/assignment3/DogShelter2.java
+/src/test/java/assignment3/Treap.java
 /out/
 /src/META-INF/
