--- src/assignment3/Replace DogShelter.java template with your own file
@@ -1,1 +1,0 @@
-
