+++ src/test/java/assignment3/Replace DogShelter.java template with your own file
@@ -1,0 +1,1 @@
+
